устранение замечаний и добавление фона
