+++ main2.py
@@ -2,41 +2,18 @@ from fastapi import FastAPI, Depends, Query, Cookie
 from fastapi.responses import HTMLResponse, JSONResponse
 from fastapi.templating import Jinja2Templates
 from fastapi.requests import Request
-from models import User, Base, Users, CustomException, CountCoffeeCups
+from models import UserRegisterModel, Base, Users, CustomException, CountCoffeeCups
 from sqlalchemy import MetaData, create_engine, select, insert, update
 from sqlalchemy.orm import sessionmaker
-from passlib.context import CryptContext
-from typing import Optional
-import jwt
+from security import *
+from fastapi.staticfiles import StaticFiles
 
 app = FastAPI()
 templates = Jinja2Templates(directory="templates")
 metadata = MetaData()
 engine = create_engine("sqlite:///database.db",echo=True)
 sess = sessionmaker(engine)
-
-# Хэширование пароля
-pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")
-
-SECRET_KEY = "qwerty"
-ALGORITHM = "HS256"
-
-
-
-
-def make_jwt_token(data):
-    return jwt.encode(data, SECRET_KEY, algorithm=ALGORITHM)
-
-
-def get_user_from_jwt_token(jwt_token: str) -> Optional[Users]:
-    try:
-        payload = jwt.decode(jwt_token, SECRET_KEY, algorithms=[ALGORITHM])
-        return payload.get("sub")
-    except:
-        return CustomException(detail="Ошибка получения JWT", status_code=400)
-
-
-
+app.mount("/static", StaticFiles(directory="static"), name="static")
 
 def create_tables():
     engine.echo = False
@@ -53,38 +30,37 @@ async def register_form(request: Request):
 
 
 @app.post("/register")
-async def register(user: User = Depends(User.as_form)):
+async def register(user: UserRegisterModel = Depends(UserRegisterModel.as_form)):
     hashed_password = pwd_context.hash(user.password)
     with sess() as session:
         query = select(Users)
         res = session.execute(query.where(Users.username == user.username)).scalar_one_or_none()
         if not res:
-            session.execute(insert(Users), {"username":user.username, "password":hashed_password, "JWT_token": ''})
+            session.execute(insert(Users), {"username":user.username, "password":hashed_password, "jwt_token": ''})
         else:
             raise CustomException(detail="Пользователь с таким именем уже существует, попробуйте другое имя", status_code=404)        
         session.commit()
 
 @app.get("/login", response_class=HTMLResponse)
-async def register_form(request: Request):
+async def login(request: Request):
     return templates.TemplateResponse("login.html", {"request": request})
 
 
 
 @app.post("/login", response_class=HTMLResponse)
-async def login(user: User = Depends(User.as_form)):
+async def login_form(user: UserRegisterModel = Depends(UserRegisterModel.as_form)):
     with sess() as session:
-        query = select(Users)
-        res = session.execute(query.where(Users.username == user.username)).scalar_one_or_none()        
+        jwt_token = make_jwt_token({"sub": user.username})
+        res = session.execute(select(Users).where(Users.username == user.username)).scalar_one_or_none()        
         if not res:
             raise CustomException(detail="Неверное имя или пароль", status_code=404)        
         # Проверяем пароль
         if not pwd_context.verify(user.password, res.password):
             raise CustomException(detail="Неверное имя или пароль", status_code=404)
-        jwt_token = make_jwt_token({"sub": user.username})
-        session.execute(update(Users).where(user.username == Users.username), {"JWT_token": jwt_token})
+        session.execute(update(Users).where(user.username == Users.username), {"jwt_token": jwt_token})
+        session.commit()
         response = JSONResponse({"message": "Успешный вход"})
         response.set_cookie("jwt", jwt_token)
-        session.commit()
         return response
 
 
@@ -96,15 +72,13 @@ async def home(request: Request):
 @app.get("/drink_coffee/", response_class=HTMLResponse)
 async def count_cup(request: Request, jwt: Optional[str] = Cookie(None)):
     user = get_user_from_jwt_token(jwt)
-    print(user)
     with sess() as session:
-        query = select(Users)
-        user_info = session.execute(query.where(user == Users.username)).scalar_one_or_none()
-        query = select(CountCoffeeCups)
-        res = session.execute(query.where(user_info.id == CountCoffeeCups.user_id)).scalar_one_or_none()
+        user_info = session.execute(select(Users).where(user == Users.username)).scalar_one_or_none()
+        res = session.execute(select(CountCoffeeCups).where(user_info.id == CountCoffeeCups.user_id)).scalar_one_or_none()
         if not res:
             session.execute(insert(CountCoffeeCups), {"user_id":user_info.id, "count_cups": 0})
-        res.count_cups += 1
-        session.execute(update(CountCoffeeCups).where(user_info.id == CountCoffeeCups.user_id), {"count_cups": res.count_cups})
+            session.commit()
+        session.execute(update(CountCoffeeCups).where(user_info.id == CountCoffeeCups.user_id).values(count_cups = CountCoffeeCups.count_cups + 1))
         session.commit()
-    return templates.TemplateResponse(name="home.html", context={"request":request, "count_cups": res.count_cups})
+        update_cups = session.execute(select(CountCoffeeCups.count_cups).where(user_info.id == CountCoffeeCups.user_id)).scalar_one()
+    return templates.TemplateResponse(name="home.html", context={"request":request, "count_cups": update_cups})
